refactor: improve typing of indoor geocoding logic
