+++ app/types/poi.ts
@@ -1,5 +1,5 @@
 export interface POI {
   name: string;
-  coordinates: number[];
+  coordinates: [number, number];
   // add any additional properties as needed
 }
